+++ src/api/tests/feature_tests/steps/tests/test_endpoint_lambda_mapping.py
@@ -127,6 +127,19 @@ def test_parse_path_read_cpm_product():
     ) == ({"product_team_id": "123", "product_id": "456"}, {}, api.readCpmProduct.index)
 
 
+def test_parse_path_search_product():
+    with api_lambda_environment_variables():
+        import api.searchCpmProduct.index
+
+        endpoint_lambda_mapping = get_endpoint_lambda_mapping()
+
+    assert parse_api_path(
+        method="GET",
+        path="ProductTeam/123/Product",
+        endpoint_lambda_mapping=endpoint_lambda_mapping,
+    ) == ({"product_team_id": "123"}, {}, api.searchCpmProduct.index)
+
+
 def test_parse_path_error():
     with pytest.raises(EndpointConfigurationError):
         parse_api_path(method="GET", path="ProductTeam/123", endpoint_lambda_mapping={})