+++ src/layers/domain/response/aws_lambda_response.py
@@ -1,5 +1,5 @@
 from http import HTTPStatus
-from typing import Literal, Optional
+from typing import Literal
 
 from pydantic import BaseModel, Field, validator
 
@@ -8,10 +8,10 @@ class AwsLambdaResponseHeaders(BaseModel):
     content_type: Literal["application/json"] = Field(
         default="application/json", alias="Content-Type"
     )
-    content_length: str = Field(alias="Content-Length", regex=r"^[1-9][0-9]*$")
+    content_length: str = Field(alias="Content-Length", regex=r"^[0-9][0-9]*$")
     version: str = Field(alias="Version", regex=r"^(null)|([1-9][0-9]*)$")
-    location: Optional[str] = Field(alias="Location")
-    host: Optional[str] = Field(alias="Host")
+    location: str = Field(default=None, alias="Location")
+    host: str = Field(default=None, alias="Host")
 
     class Config:
         allow_population_by_field_name = True
@@ -22,10 +22,10 @@ class AwsLambdaResponseHeaders(BaseModel):
 
 class AwsLambdaResponse(BaseModel):
     statusCode: HTTPStatus
-    body: str = Field(min_length=1)
+    body: str = Field(min_length=0, default="")
     version: None | str = Field(exclude=True)
     location: None | str = Field(exclude=True, default=None)
-    headers: Optional[AwsLambdaResponseHeaders] = None
+    headers: AwsLambdaResponseHeaders = None
 
     @validator("headers", always=True)
     def generate_response_headers(cls, headers, values):